refactor: extract shared patch helpers
