+++ src/core/loaders.ts
@@ -1,5 +1,4 @@
 import {
-  getFiletypeFromFileName,
   parseDiffFromFile,
   parsePatchFiles,
   type FileContents,
@@ -9,9 +8,11 @@ import { createTwoFilesPatch } from "diff";
 import fs from "node:fs";
 import { join, resolve as resolvePath } from "node:path";
 import { findAgentFileContext, loadAgentContext } from "./agent";
-import { createSkippedBinaryMetadata, isProbablyBinaryFile, patchLooksBinary } from "./binary";
-import { normalizeDiffMetadataPaths, normalizeDiffPath } from "./diffPaths";
-import { runGitUntrackedFileDiffText } from "./git";
+import { createSkippedBinaryMetadata, isProbablyBinaryFile } from "./binary";
+import { normalizeUntrackedPatchHeaders, runGitUntrackedFileDiffText } from "./git";
+import { splitPatchIntoFileChunks, findPatchChunk } from "./patch/chunks";
+import { buildDiffFile, createSkippedLargeMetadata } from "./diffFile";
+import { normalizePatchText } from "./patch/normalize";
 import { createUnsupportedVcsOperationError, getVcsAdapter, operationFromInput } from "./vcs";
 import type {
   AppBootstrap,
@@ -38,497 +39,6 @@ const LARGE_DIFF_FILE_SNIFF_BYTES = 256 * 1024;
 /** Return the final path segment for display-oriented labels. */
 function basename(path: string) {
   return path.split(/[\\/]/).filter(Boolean).pop() ?? path;
-}
-
-/** Remove git-style a/ and b/ prefixes before matching diff paths. */
-function stripPrefixes(path: string) {
-  return path.replace(/^[ab]\//, "");
-}
-
-/** Remove terminal escape sequences so Git-colored pager input still parses as plain patch text. */
-function stripTerminalControl(text: string) {
-  return text
-    .replace(/\x1bP[\s\S]*?\x1b\\/g, "")
-    .replace(/\x1b\][\s\S]*?(?:\x07|\x1b\\)/g, "")
-    .replace(/\x1b\[[0-?]*[ -/]*[@-~]/g, "")
-    .replace(/\x1b[@-_]/g, "");
-}
-
-/**
- * Strip `git log -p` / `git show -p` commit metadata so the surviving text
- * is a plain patch stream that `@pierre/diffs` can parse without spamming
- * `parseLineType: Invalid firstChar` warnings on every commit boundary.
- *
- * Each commit in `git log -p` looks like:
- *
- * ```
- * commit <sha>[ (refs)]
- * Author: ...
- * Date:   ...
- *
- *     <commit message>
- *
- * diff --git a/foo b/foo
- * ...
- * ```
- *
- * Lines from `commit ` through the first patch header (`diff --git `,
- * `--- `, or `+++ `) are dropped. Hunk-body lines always start with
- * `+`, `-`, ` ` or `\`, so a real context line that begins with the word
- * "commit" is unaffected (its leading space prevents the regex match).
- *
- * Returns the input unchanged when no `commit <sha>` boundary is present,
- * keeping the regular patch path zero-cost.
- */
-export function stripGitLogMetadata(text: string) {
-  // Hex range up to 64 covers both SHA-1 (40) and SHA-256 (64) repos.
-  const COMMIT_BOUNDARY = /^commit [0-9a-f]{4,64}(?: |$)/m;
-  if (!COMMIT_BOUNDARY.test(text)) {
-    return text;
-  }
-
-  const lines = text.split("\n");
-  const out: string[] = [];
-  let inHeader = false;
-
-  for (const line of lines) {
-    if (COMMIT_BOUNDARY.test(line)) {
-      inHeader = true;
-      continue;
-    }
-    if (inHeader) {
-      // The header section ends at the first patch line. `diff --git `
-      // is the canonical Git start; `--- `/`+++ ` cover unified-diff
-      // input where someone synthesised log output without it.
-      if (line.startsWith("diff --git ") || line.startsWith("--- ") || line.startsWith("+++ ")) {
-        inHeader = false;
-        out.push(line);
-      }
-      continue;
-    }
-    out.push(line);
-  }
-
-  return out.join("\n");
-}
-
-/** Split a multi-file patch into per-file chunks so each diff file keeps its original patch text. */
-function splitPatchIntoFileChunks(rawPatch: string) {
-  const patch = rawPatch.replaceAll("\r\n", "\n");
-  const lines = patch.split("\n");
-  const chunks: string[] = [];
-  let current: string[] = [];
-  const hasGitHeaders = lines.some((line) => line.startsWith("diff --git "));
-
-  const flush = () => {
-    if (current.length > 0) {
-      chunks.push(`${current.join("\n").trimEnd()}\n`);
-      current = [];
-    }
-  };
-
-  for (let index = 0; index < lines.length; index += 1) {
-    const line = lines[index]!;
-
-    if (hasGitHeaders && line.startsWith("diff --git ")) {
-      flush();
-      current.push(line);
-      continue;
-    }
-
-    if (!hasGitHeaders && line.startsWith("--- ") && lines[index + 1]?.startsWith("+++ ")) {
-      flush();
-      current.push(line);
-      current.push(lines[index + 1]!);
-      index += 1;
-      continue;
-    }
-
-    if (current.length > 0) {
-      current.push(line);
-    }
-  }
-
-  flush();
-  return chunks;
-}
-
-/** Count visible additions and deletions from parsed diff metadata. */
-function countDiffStats(metadata: FileDiffMetadata) {
-  let additions = 0;
-  let deletions = 0;
-
-  for (const hunk of metadata.hunks) {
-    for (const content of hunk.hunkContent) {
-      if (content.type === "change") {
-        additions += content.additions;
-        deletions += content.deletions;
-      }
-    }
-  }
-
-  return { additions, deletions };
-}
-
-/** Recover the original patch chunk for one parsed file, preferring index order before path matching. */
-function findPatchChunk(metadata: FileDiffMetadata, chunks: string[], index: number) {
-  const byIndex = chunks[index];
-  if (byIndex) {
-    return byIndex;
-  }
-
-  return (
-    chunks.find((chunk) =>
-      [metadata.name, metadata.prevName]
-        .map(normalizeDiffPath)
-        .filter((value): value is string => Boolean(value))
-        .map(stripPrefixes)
-        .some(
-          (path) =>
-            chunk.includes(`a/${path}`) || chunk.includes(`b/${path}`) || chunk.includes(path),
-        ),
-    ) ?? ""
-  );
-}
-
-interface BuildDiffFileOptions {
-  isUntracked?: boolean;
-  previousPath?: string;
-  isBinary?: boolean;
-  isTooLarge?: boolean;
-  stats?: DiffFile["stats"];
-  statsTruncated?: boolean;
-}
-
-/** Build the normalized per-file model used by the UI regardless of input mode. */
-function buildDiffFile(
-  metadata: FileDiffMetadata,
-  patch: string,
-  index: number,
-  sourcePrefix: string,
-  agentContext: AgentContext | null,
-  {
-    isUntracked,
-    previousPath,
-    isBinary,
-    isTooLarge,
-    stats,
-    statsTruncated,
-  }: BuildDiffFileOptions = {},
-): DiffFile {
-  const normalizedMetadata = normalizeDiffMetadataPaths(metadata);
-  const path = normalizedMetadata.name;
-  const resolvedPreviousPath = normalizeDiffPath(previousPath) ?? normalizedMetadata.prevName;
-
-  return {
-    id: `${sourcePrefix}:${index}:${path}`,
-    path,
-    previousPath: resolvedPreviousPath,
-    patch,
-    language: getFiletypeFromFileName(path) ?? undefined,
-    stats: stats ?? countDiffStats(normalizedMetadata),
-    metadata: normalizedMetadata,
-    agent: findAgentFileContext(agentContext, path, resolvedPreviousPath),
-    isUntracked,
-    isBinary: isBinary ?? patchLooksBinary(patch),
-    isTooLarge,
-    statsTruncated,
-  };
-}
-
-/**
- * Re-add Git's `a/` and `b/` path prefixes to patch headers when stdin came from a
- * `git diff` that was emitted with `diff.noprefix=true` (or otherwise stripped prefixes).
- *
- * `@pierre/diffs` requires `a/` and `b/` on `diff --git`, `---`, and `+++` lines and throws
- * a `TypeError` on the first noprefix header, which leaves the review with zero files. The
- * git-backed paths force `diff.noprefix=false` when they invoke git internally; this helper
- * covers the patch path (`hunk patch`, `hunk pager`) where the input was produced by an
- * outer `git` process we do not control.
- *
- * The rewrite is scoped to header lines only: once the `+++ ` line has been emitted for a
- * block we clear the flag so a deleted line whose content starts with `-- ` (e.g. a removed
- * SQL/Lua/Haskell comment, which becomes `--- foo` on disk) is not mistaken for a file
- * header inside the hunk body.
- */
-type GitHeaderRewriteMode = "add" | "strip";
-
-function normalizeGitPatchPrefixes(patchText: string) {
-  if (!patchText.includes("diff --git ")) {
-    return patchText;
-  }
-
-  const lines = patchText.split("\n");
-  const normalizedLines: string[] = [];
-  let blockLines: string[] = [];
-
-  const flushBlock = () => {
-    if (blockLines.length === 0) {
-      return;
-    }
-
-    for (const line of rewriteGitPatchBlock(blockLines)) {
-      normalizedLines.push(line);
-    }
-    blockLines = [];
-  };
-
-  for (const line of lines) {
-    if (line.startsWith("diff --git ")) {
-      flushBlock();
-      blockLines.push(line);
-      continue;
-    }
-
-    if (blockLines.length > 0) {
-      blockLines.push(line);
-    } else {
-      normalizedLines.push(line);
-    }
-  }
-
-  flushBlock();
-  return normalizedLines.join("\n");
-}
-
-/** Rewrite one `diff --git` block, keeping file-header rewrites out of hunk bodies. */
-function rewriteGitPatchBlock(blockLines: string[]) {
-  const firstLine = blockLines[0];
-  if (!firstLine?.startsWith("diff --git ")) {
-    return blockLines;
-  }
-
-  const result = rewriteGitDiffHeader(firstLine, blockLines);
-  let blockRewriteMode = result.rewriteMode;
-
-  const rewrittenLines = [result.line];
-
-  for (const line of blockLines.slice(1)) {
-    if (blockRewriteMode && line.startsWith("--- ")) {
-      rewrittenLines.push(rewriteUnifiedFileLine(line, "--- ", "a/", blockRewriteMode));
-      continue;
-    }
-
-    if (blockRewriteMode && line.startsWith("+++ ")) {
-      const rewriteMode = blockRewriteMode;
-      blockRewriteMode = null;
-      rewrittenLines.push(rewriteUnifiedFileLine(line, "+++ ", "b/", rewriteMode));
-      continue;
-    }
-
-    rewrittenLines.push(line);
-  }
-
-  return rewrittenLines;
-}
-
-/** Detect prefixed/noprefix `diff --git` lines and rewrite them into Pierre's `a/X b/Y` form. */
-function rewriteGitDiffHeader(
-  line: string,
-  blockLines: string[],
-): {
-  line: string;
-  rewriteMode: GitHeaderRewriteMode | null;
-} {
-  const rest = line.slice("diff --git ".length).trimEnd();
-
-  const quotedMatch = rest.match(/^"((?:\\.|[^"\\])*)" "((?:\\.|[^"\\])*)"$/);
-  if (quotedMatch) {
-    const [, oldPath = "", newPath = ""] = quotedMatch;
-    const pair = canonicalizeGitPathPair(oldPath, newPath, blockLines);
-    // Pierre's git header parser does not currently handle the quoted `"a/..." "b/..."`
-    // form, so canonicalize quoted paths to the unquoted form even when prefixes exist.
-    return { line: `diff --git ${pair.oldPath} ${pair.newPath}`, rewriteMode: pair.rewriteMode };
-  }
-
-  const tokens = rest.split(" ");
-
-  if (tokens.length >= 2 && tokens.length % 2 === 0) {
-    const half = tokens.length / 2;
-    const firstHalf = tokens.slice(0, half).join(" ");
-    const secondHalf = tokens.slice(half).join(" ");
-    const knownPair = canonicalizeKnownGitPathPair(firstHalf, secondHalf, blockLines);
-
-    if (knownPair?.changed) {
-      return {
-        line: `diff --git ${knownPair.oldPath} ${knownPair.newPath}`,
-        rewriteMode: knownPair.rewriteMode,
-      };
-    }
-
-    // Already prefixed: `a/X b/Y` (covers single-token and equally split multi-token paths).
-    if (knownPair?.isCanonical) {
-      return { line, rewriteMode: null };
-    }
-
-    // Non-rename noprefix: identical halves regardless of whether the path contains spaces.
-    if (firstHalf === secondHalf && firstHalf.length > 0) {
-      return { line: `diff --git a/${firstHalf} b/${secondHalf}`, rewriteMode: "add" };
-    }
-  }
-
-  // Two-token rename without prefix and without spaces in either path.
-  if (tokens.length === 2 && tokens[0] && tokens[1]) {
-    return { line: `diff --git a/${tokens[0]} b/${tokens[1]}`, rewriteMode: "add" };
-  }
-
-  // Genuinely ambiguous (rename with spaces and no quoting). Leave untouched and let the
-  // parser surface the existing failure rather than guess at the path split.
-  return { line, rewriteMode: null };
-}
-
-const GIT_MNEMONIC_PREFIXES = new Set(["c", "i", "o", "w", "1", "2"]);
-
-/** Return one Git mnemonic side prefix from a path, if present. */
-function splitGitMnemonicPrefix(path: string) {
-  const [prefix, ...rest] = path.split("/");
-  if (!prefix || rest.length === 0 || !GIT_MNEMONIC_PREFIXES.has(prefix)) {
-    return null;
-  }
-
-  return { prefix, path: rest.join("/") };
-}
-
-/** Remove Git's outer quotes from one path-like metadata value. */
-function stripGitPathQuotes(path: string) {
-  return path.match(/^"((?:\\.|[^"\\])*)"$/)?.[1] ?? path;
-}
-
-/** Return rename metadata, which Git writes without mnemonic side prefixes. */
-function findRenameMetadata(blockLines: string[]) {
-  const oldPath = blockLines.find((line) => line.startsWith("rename from "));
-  const newPath = blockLines.find((line) => line.startsWith("rename to "));
-
-  if (!oldPath || !newPath) {
-    return null;
-  }
-
-  return {
-    oldPath: stripGitPathQuotes(oldPath.slice("rename from ".length)),
-    newPath: stripGitPathQuotes(newPath.slice("rename to ".length)),
-  };
-}
-
-/** Return a path with the expected Git side prefix while avoiding double-prefixing. */
-function withGitPrefix(path: string, prefix: "a/" | "b/") {
-  return path.startsWith(prefix) ? path : `${prefix}${path}`;
-}
-
-/** Decide whether a mnemonic-looking path pair is real mnemonic output or a noprefix rename. */
-function shouldStripMnemonicPair(oldPath: string, newPath: string, blockLines: string[]) {
-  const oldMnemonic = splitGitMnemonicPrefix(oldPath);
-  const newMnemonic = splitGitMnemonicPrefix(newPath);
-
-  if (!oldMnemonic || !newMnemonic || oldMnemonic.prefix === newMnemonic.prefix) {
-    return null;
-  }
-
-  const rename = findRenameMetadata(blockLines);
-  if (!rename) {
-    return true;
-  }
-
-  if (rename.oldPath === oldPath && rename.newPath === newPath) {
-    return false;
-  }
-
-  if (rename.oldPath === oldMnemonic.path && rename.newPath === newMnemonic.path) {
-    return true;
-  }
-
-  return true;
-}
-
-/** Convert already-prefixed or mnemonic-prefixed path pairs into Pierre's canonical shape. */
-function canonicalizeKnownGitPathPair(oldPath: string, newPath: string, blockLines: string[]) {
-  const oldMnemonic = splitGitMnemonicPrefix(oldPath);
-  const newMnemonic = splitGitMnemonicPrefix(newPath);
-  const isCanonical = oldPath.startsWith("a/") && newPath.startsWith("b/");
-
-  if (isCanonical) {
-    return { oldPath, newPath, rewriteMode: "add" as const, changed: false, isCanonical: true };
-  }
-
-  if (oldMnemonic && newMnemonic && shouldStripMnemonicPair(oldPath, newPath, blockLines)) {
-    return {
-      oldPath: `a/${oldMnemonic.path}`,
-      newPath: `b/${newMnemonic.path}`,
-      rewriteMode: "strip" as const,
-      changed: true,
-      isCanonical: false,
-    };
-  }
-
-  return null;
-}
-
-/** Convert one quoted `diff --git` path pair into Pierre's canonical side-prefix shape. */
-function canonicalizeGitPathPair(oldPath: string, newPath: string, blockLines: string[]) {
-  return (
-    canonicalizeKnownGitPathPair(oldPath, newPath, blockLines) ?? {
-      oldPath: withGitPrefix(oldPath, "a/"),
-      newPath: withGitPrefix(newPath, "b/"),
-      rewriteMode: "add" as const,
-      changed: true,
-      isCanonical: false,
-    }
-  );
-}
-
-/** Insert the canonical `a/` or `b/` prefix on a unified-diff header that is missing it. */
-function rewriteUnifiedFileLine(
-  line: string,
-  marker: "--- " | "+++ ",
-  prefix: "a/" | "b/",
-  mode: GitHeaderRewriteMode,
-) {
-  const path = line.slice(marker.length);
-  const quotedPath = path.match(/^"((?:\\.|[^"\\])*)"(.*)$/);
-  const pathName = quotedPath?.[1] ?? path;
-  const suffix = quotedPath?.[2] ?? "";
-
-  if (pathName === "/dev/null" || pathName.startsWith("/dev/null\t")) {
-    return line;
-  }
-
-  const normalizedPath =
-    mode === "strip" ? (splitGitMnemonicPrefix(pathName)?.path ?? pathName) : pathName;
-
-  return `${marker}${withGitPrefix(normalizedPath, prefix)}${suffix}`;
-}
-
-/** Escape only the filename characters that break unified-diff header parsing. */
-function escapeUntrackedPatchPath(path: string) {
-  return path
-    .replaceAll("\\", "\\\\")
-    .replaceAll("\t", "\\t")
-    .replaceAll("\n", "\\n")
-    .replaceAll("\r", "\\r");
-}
-
-/** Rewrite Git's quoted untracked-file headers into parser-friendly paths. */
-function normalizeUntrackedPatchHeaders(patchText: string, filePath: string) {
-  const safePath = escapeUntrackedPatchPath(filePath);
-
-  return patchText
-    .replaceAll("\r\n", "\n")
-    .split("\n")
-    .map((line) => {
-      if (line.startsWith("diff --git ")) {
-        return `diff --git a/${safePath} b/${safePath}`;
-      }
-
-      if (line.startsWith("+++ ")) {
-        return `+++ b/${safePath}`;
-      }
-
-      if (line.startsWith("Binary files /dev/null and ")) {
-        return `Binary files /dev/null and b/${safePath} differ`;
-      }
-
-      return line;
-    })
-    .join("\n");
 }
 
 interface CountedLines {
@@ -603,24 +113,6 @@ function inspectLargeUntrackedFile(repoRoot: string, filePath: string): LargeUnt
     shouldSkip,
     stats: shouldSkip ? { additions: counted.lines, deletions: 0 } : undefined,
     statsTruncated: shouldSkip ? !counted.complete : undefined,
-  };
-}
-
-/** Build placeholder metadata for a file whose full diff would be too expensive. */
-function createSkippedLargeMetadata(
-  filePath: string,
-  type: FileDiffMetadata["type"],
-): FileDiffMetadata {
-  return {
-    name: filePath,
-    type,
-    hunks: [],
-    splitLineCount: 0,
-    unifiedLineCount: 0,
-    isPartial: true,
-    additionLines: [],
-    deletionLines: [],
-    cacheKey: `${filePath}:large-diff-skipped`,
   };
 }
 
@@ -739,9 +231,7 @@ function normalizePatchChangeset(
   sourceLabel: string,
   agentContext: AgentContext | null,
 ): Changeset {
-  const normalizedPatchText = normalizeGitPatchPrefixes(
-    stripGitLogMetadata(stripTerminalControl(patchText.replaceAll("\r\n", "\n"))),
-  );
+  const normalizedPatchText = normalizePatchText(patchText);
 
   let parsedPatches: ReturnType<typeof parsePatchFiles>;
   try {
