+++ src/core/patch/chunks.ts
@@ -1,0 +1,69 @@
+import type { FileDiffMetadata } from "@pierre/diffs";
+import { normalizeDiffPath } from "../diffPaths";
+
+/** Remove git-style a/ and b/ prefixes before matching diff paths. */
+function stripPrefixes(path: string) {
+  return path.replace(/^[ab]\//, "");
+}
+
+/** Split a multi-file patch into per-file chunks so each diff file keeps its original patch text. */
+export function splitPatchIntoFileChunks(rawPatch: string) {
+  const patch = rawPatch.replaceAll("\r\n", "\n");
+  const lines = patch.split("\n");
+  const chunks: string[] = [];
+  let current: string[] = [];
+  const hasGitHeaders = lines.some((line) => line.startsWith("diff --git "));
+
+  const flush = () => {
+    if (current.length > 0) {
+      chunks.push(`${current.join("\n").trimEnd()}\n`);
+      current = [];
+    }
+  };
+
+  for (let index = 0; index < lines.length; index += 1) {
+    const line = lines[index]!;
+
+    if (hasGitHeaders && line.startsWith("diff --git ")) {
+      flush();
+      current.push(line);
+      continue;
+    }
+
+    if (!hasGitHeaders && line.startsWith("--- ") && lines[index + 1]?.startsWith("+++ ")) {
+      flush();
+      current.push(line);
+      current.push(lines[index + 1]!);
+      index += 1;
+      continue;
+    }
+
+    if (current.length > 0) {
+      current.push(line);
+    }
+  }
+
+  flush();
+  return chunks;
+}
+
+/** Recover the original patch chunk for one parsed file, preferring index order before path matching. */
+export function findPatchChunk(metadata: FileDiffMetadata, chunks: string[], index: number) {
+  const byIndex = chunks[index];
+  if (byIndex) {
+    return byIndex;
+  }
+
+  return (
+    chunks.find((chunk) =>
+      [metadata.name, metadata.prevName]
+        .map(normalizeDiffPath)
+        .filter((value): value is string => Boolean(value))
+        .map(stripPrefixes)
+        .some(
+          (path) =>
+            chunk.includes(`a/${path}`) || chunk.includes(`b/${path}`) || chunk.includes(path),
+        ),
+    ) ?? ""
+  );
+}
